Backend/Microservices/user_service/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
pydantic
alembic
python-dotenv
passlib
bcrypt
